requirements.txt
future==0.16.0
imgurpython==1.1.7
oauthlib==2.0.1
Pygments==2.1.3
python-twitter==2.2
requests==2.12.4
requests-oauthlib==0.7.0
urwid==1.3.1
